requirements.txt
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
pdf2image==1.16.0
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
tqdm==4.64.1
